Move validateMessageAndAddEmojiIfOk() to utils.js
